Ignore empty dtoverlay coming from target state
Before v16, the supervisor would incorrectly interpret an empty dtoverlay in
config.txt (i.e. a line with `dtoverlay=`) as a proper variable and
create it as config var if found on config.txt before provisioning.
With the latest supervisor, an empty dtoverlay on config.txt is
interpreted in the [correct way](https://www.raspberrypi.com/documentation/computers/config_txt.html#dtoverlay), but
that makes an empty dtoverlay on the target state meaningless, leading
to looping behavior as the current and target state won't match.

This PR modifies the pre-processing function to filter out empty array
variables before the comparison, to prevent this sort of looping
behavior. This allows empty dtoverlay to be ignored.

Change-type: patch