Gracefully handle multiple reboot/shutdown requests
Since HTTP's server.close() is async, there is a slim chance
for two instances of /v1/reboot or /v1/shutdown to be processed.
If the server is already closed when server.close() is called,
the call throws ERR_SERVER_NOT_RUNNING
which doesn't need to be surfaced to the user. This change
only allows one server.close() attempt to occur at a time.

Change-type: patch
Signed-off-by: Christina Ying Wang <christina@balena.io>