Update io-ts to 2.2.20, io-ts-reporters to 2.0.1, fp-ts to 2.16.5
TODO: Upgrading io-ts to 2.2.21 will result in some unit tests for
io-ts type validation breaking, due to a change in t.record's handling
of decoding unknown objects. In <2.2.21, a record with a custom key type
will decode an unknown object to Left if it has a key is not of the custom
type. Whereas in io-ts@2.2.21, t.record will strip the invalid key and
decode to Right. This is a breaking change for us.

Change-type: patch
Signed-off-by: Christina Ying Wang <christina@balena.io>