Uninstall containers using force remove
When uninstalling a container for a restart from the API on when
moving between releases, this prefers to force remove the container
rather than stop then remove. This has the same effect in the end, but
force remove is an atomic call, meaning if the supervisor is killed
between the stop and remove, the engine will proceed with the operation
and the supervisor will know to recover from the target state after the
restart.

Change-type: patch
Relates-to: #2257