Support protected field get/set/remove in ConfigJsonConfigBackend
Protected fields are defined as object fields in config.json
where the Supervisor only manages some of the fields, while
others are unmanaged (set by other OS services or preloaded
by user). The special handling for protected fields ensures
that the Supervisor does not return unmanaged fields when
getting a protected key, and does not overwrite unmanaged fields
when setting or removing.

Signed-off-by: Christina Ying Wang <christina@balena.io>