Add PowerFanConfig config backend
This config backend uses ConfigJsonConfigBackend to write to
managed fields in the "os" key, in order to set power and fan
configs. In config.json, the expected managed schema is:
```
{
  os: {
    power: {
      mode: string
    },
    fan: {
      profile: string
    }
  }
}
```

There may be other keys in os which are not managed by the Supervisor,
so the Supervisor should not read or write to them.

After the Supervisor writes to config.json, host services os-power-mode
and os-fan-profile pick up the changes, on reboot in the former's case
and at runtime in the latter's case. The changes are applied by the host
services, which the Supervisor does not manage aside from streaming
their service logs to the dashboard (see next commit).

Change-type: minor
Signed-off-by: Christina Ying Wang <christina@balena.io>