Take all locks through updateLock.takeLock
updateLock.lock disposes all locks as a final step regardless of
locking success. It also tries to take the lock for every service
under an appId, even if Supervisor locks are already taken for that
service from another context (such as from updateLock.takeLock
which is a state engine step). This latter case can result in
UpdatesLockedError where both the lock already present and the
locking attempt originate from the Supervisor, which is undesirable
as this will lead to updateLock.lock disposing all locks.

The solution is this is for any lock-taking mechanism to go through
updateLock.takeLock, which only takes Supervisor locks where not already
taken by the Supervisor. This commit splits out the lock-taking functionality
into a separate function, updateLock.takeLockFor, which takes the locks and
returns the process lock release function, allowing the callee to control
when process locks should be released. This is necessary as updateLock.lock
can call takeLockFor for multiple appIds, and process locks should only
be released after all resources in scope are locked.

Change-type: patch
Signed-off-by: Christina Ying Wang <christina@balena.io>