Reply to `/v1/purge` before restoring the state
Purging data is already a long operation and waiting the target state to
be restored before sending the response can cause the request to
timeout. With `reboot-required`, the response will never be sent
because a reboot will happen when restoring the state.

This change ensures the API response happens once purge is completed.

Change-type: patch