Add `helios` service to composition
Helios is the experimental supervisor replacement. The service acts as
a proxy for the existing supervisor and will slowly take over supervisor features

This also adds a `helios-api` service to listen on port 48484

While this should not break any of the interfaces, it is an important
change so it has a new major version.

Change-type: major