Add firewall configurations for the override port
When the override port is set, the firewall is configured to only allow
traffic on `supervisor0` interface for that port. In that case a new
chain `BALENA-SUPERVISOR` is added to `DOCKER-USER`, that chain blocks
access to the regular listen port from any interface other than
resin-vpn when in non-local-mode and allows any traffic on local-mode.

Change-type: patch